[fixed] Area attacks hitting target multiple times [fixed] Compiler warnings
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@1116 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/objects/creature/CreatureObjectImplementation.cpp
@@ -4639,9 +4639,9 @@ void CreatureObjectImplementation::onRegenerateHAM() {
 	if (isMeditating()) {
 		int meditateMod = getSkillMod("meditate");
 		float meditateBonus = 1 + ((float)meditateMod / 100);
-		healthTick *= meditateBonus;
-		actionTick *= meditateBonus;
-		mindTick *= meditateBonus;
+		healthTick *= (int)meditateBonus;
+		actionTick *= (int)meditateBonus;
+		mindTick *= (int)meditateBonus;
 		doMeditateHeals();
 	}
